pyproject.toml
[project]
name = "artmind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cattrs>=24.1.3",
    "flask>=3.1.0",
    "flask-restful>=0.3.10",
    "loguru>=0.7.3",
    "ollama>=0.4.7",
    "openai>=1.70.0",
    "pyyaml>=6.0.2",
    "streamlit>=1.44.0",
    "watchdog>=6.0.0",
]
